backend/requirements.txt
attrs==19.3.0
audio-metadata==0.11.1
bidict==0.21.2
bitstruct==8.11.1
certifi==2021.5.30
chardet==4.0.0
click==6.7
coverage==5.5
decorator==4.4.2
deprecation==2.1.0
filetype==1.0.7
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
itsdangerous==0.24
Jinja2==2.11.2
Js2Py==0.71
MarkupSafe==1.1.1
more-itertools==8.8.0
numpy==1.20.3
packaging==20.9
pendulum==2.1.2
Pillow==8.2.0
pprintpp==0.4.0
proglog==0.1.9
pyjsparser==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.8
tbm-utils==2.6.0
toml==0.10.2
tqdm==4.61.0
tzlocal==2.1
urllib3==1.26.5
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2021.6.6
ffprobe==0.5
ffmpeg==1.4
psycopg2==2.8.6
librosa